proxy as middleware and jwt verification
